Split up enums to separate files
