+++ MapboxDirections/MBRouteStep.swift
@@ -1,551 +1,5 @@
 import Polyline
 
-/**
- A `TransportType` specifies the mode of transportation used for part of a route.
- */
-@objc(MBTransportType)
-public enum TransportType: Int, CustomStringConvertible, Codable {
-    // Possible transport types when the `profileIdentifier` is `MBDirectionsProfileIdentifierAutomobile` or `MBDirectionsProfileIdentifierAutomobileAvoidingTraffic`
-    
-    /**
-     The step does not have a particular transport type associated with it.
-     
-     This transport type is used as a workaround for bridging to Objective-C which does not support nullable enumeration-typed values.
-     */
-    case none
-    
-    /**
-     The route requires the user to drive or ride a car, truck, or motorcycle.
-     
-     This is the usual transport type when the `profileIdentifier` is `MBDirectionsProfileIdentifierAutomobile` or `MBDirectionsProfileIdentifierAutomobileAvoidingTraffic`.
-     */
-    case automobile // automobile
-    
-    /**
-     The route requires the user to board a ferry.
-     
-     The user should verify that the ferry is in operation. For driving and cycling directions, the user should also verify that his or her vehicle is permitted onboard the ferry.
-     */
-    case ferry // automobile, walking, cycling
-    
-    /**
-     The route requires the user to cross a movable bridge.
-     
-     The user may need to wait for the movable bridge to become passable before continuing.
-     */
-    case movableBridge // automobile, cycling
-    
-    /**
-     The route becomes impassable at this point.
-     
-     You should not encounter this transport type under normal circumstances.
-     */
-    case inaccessible // automobile, walking, cycling
-    
-    // Possible transport types when the `profileIdentifier` is `MBDirectionsProfileIdentifierWalking`
-    
-    /**
-     The route requires the user to walk.
-     
-     This is the usual transport type when the `profileIdentifier` is `MBDirectionsProfileIdentifierWalking`. For cycling directions, this value indicates that the user is expected to dismount.
-     */
-    case walking // walking, cycling
-    
-    // Possible transport types when the `profileIdentifier` is `MBDirectionsProfileIdentifierCycling`
-    
-    /**
-     The route requires the user to ride a bicycle.
-     
-     This is the usual transport type when the `profileIdentifier` is `MBDirectionsProfileIdentifierCycling`.
-     */
-    case cycling // cycling
-    
-    /**
-     The route requires the user to board a train.
-     
-     The user should consult the train’s timetable. For cycling directions, the user should also verify that bicycles are permitted onboard the train.
-     */
-    case train // cycling
-    
-    public init?(description: String) {
-        let type: TransportType
-        switch description {
-        case "none":
-            type = .none
-        case "driving":
-            type = .automobile
-        case "ferry":
-            type = .ferry
-        case "moveable bridge":
-            type = .movableBridge
-        case "unaccessible":
-            type = .inaccessible
-        case "walking":
-            type = .walking
-        case "cycling":
-            type = .cycling
-        case "train":
-            type = .train
-        default:
-            return nil
-        }
-        self.init(rawValue: type.rawValue)
-    }
-    
-    public var description: String {
-        switch self {
-        case .none:
-            return "none"
-        case .automobile:
-            return "driving"
-        case .ferry:
-            return "ferry"
-        case .movableBridge:
-            return "moveable bridge"
-        case .inaccessible:
-            return "unaccessible"
-        case .walking:
-            return "walking"
-        case .cycling:
-            return "cycling"
-        case .train:
-            return "train"
-        }
-    }
-}
-
-/**
- A `ManeuverType` specifies the type of maneuver required to complete the route step. You can pair a maneuver type with a `ManeuverDirection` to choose an appropriate visual or voice prompt to present the user.
- 
- In Swift, you can use pattern matching with a single switch statement on a tuple containing the maneuver type and maneuver direction to avoid a complex series of if-else-if statements or switch statements.
- */
-@objc(MBManeuverType)
-public enum ManeuverType: Int, CustomStringConvertible, Codable {
-    /**
-     The step does not have a particular maneuver type associated with it.
-     
-     This maneuver type is used as a workaround for bridging to Objective-C which does not support nullable enumeration-typed values.
-     */
-    case none
-    
-    /**
-     The step requires the user to depart from a waypoint.
-     
-     If the waypoint is some distance away from the nearest road, the maneuver direction indicates the direction the user must turn upon reaching the road.
-     */
-    case depart
-    
-    /**
-     The step requires the user to turn.
-     
-     The maneuver direction indicates the direction in which the user must turn relative to the current direction of travel. The exit index indicates the number of intersections, large or small, from the previous maneuver up to and including the intersection at which the user must turn.
-     */
-    case turn
-    
-    /**
-     The step requires the user to continue after a turn.
-     */
-    case `continue`
-    
-    /**
-     The step requires the user to continue on the current road as it changes names.
-     
-     The step’s name contains the road’s new name. To get the road’s old name, use the previous step’s name.
-     */
-    case passNameChange
-    
-    /**
-     The step requires the user to merge onto another road.
-     
-     The maneuver direction indicates the side from which the other road approaches the intersection relative to the user.
-     */
-    case merge
-    
-    /**
-     The step requires the user to take a entrance ramp (slip road) onto a highway.
-     */
-    case takeOnRamp
-    
-    /**
-     The step requires the user to take an exit ramp (slip road) off a highway.
-     
-     The maneuver direction indicates the side of the highway from which the user must exit. The exit index indicates the number of highway exits from the previous maneuver up to and including the exit that the user must take.
-     */
-    case takeOffRamp
-    
-    /**
-     The step requires the user to choose a fork at a Y-shaped fork in the road.
-     
-     The maneuver direction indicates which fork to take.
-     */
-    case reachFork
-    
-    /**
-     The step requires the user to turn at either a T-shaped three-way intersection or a sharp bend in the road where the road also changes names.
-     
-     This maneuver type is called out separately so that the user may be able to proceed more confidently, without fear of having overshot the turn. If this distinction is unimportant to you, you may treat the maneuver as an ordinary `turn`.
-     */
-    case reachEnd
-    
-    /**
-     The step requires the user to get into a specific lane in order to continue along the current road.
-     
-     The maneuver direction is set to `straightAhead`. Each of the first intersection’s usable approach lanes also has an indication of `straightAhead`. A maneuver in a different direction would instead have a maneuver type of `turn`.
-     
-     This maneuver type is called out separately so that the application can present the user with lane guidance based on the first element in the `intersections` property. If lane guidance is unimportant to you, you may treat the maneuver as an ordinary `continue` or ignore it.
-     */
-    case useLane
-     
-    /**
-     The step requires the user to enter and traverse a roundabout (traffic circle or rotary).
-     
-     The step has no name, but the exit name is the name of the road to take to exit the roundabout. The exit index indicates the number of roundabout exits up to and including the exit to take.
-     
-     If `RouteOptions.includesExitRoundaboutManeuver` is set to `true`, this step is followed by an `.exitRoundabout` maneuver. Otherwise, this step represents the entire roundabout maneuver, from the entrance to the exit.
-     */
-    case takeRoundabout
-    
-    /**
-     The step requires the user to enter and traverse a large, named roundabout (traffic circle or rotary).
-     
-     The step’s name is the name of the roundabout. The exit name is the name of the road to take to exit the roundabout. The exit index indicates the number of rotary exits up to and including the exit that the user must take.
-     
-      If `RouteOptions.includesExitRoundaboutManeuver` is set to `true`, this step is followed by an `.exitRotary` maneuver. Otherwise, this step represents the entire roundabout maneuver, from the entrance to the exit.
-     */
-    case takeRotary
-    
-    /**
-     The step requires the user to enter and exit a roundabout (traffic circle or rotary) that is compact enough to constitute a single intersection.
-     
-     The step’s name is the name of the road to take after exiting the roundabout. This maneuver type is called out separately because the user may perceive the roundabout as an ordinary intersection with an island in the middle. If this distinction is unimportant to you, you may treat the maneuver as either an ordinary `turn` or as a `takeRoundabout`.
-     */
-    case turnAtRoundabout
-    
-    /**
-     The step requires the user to exit a roundabout (traffic circle or rotary).
-     
-     This maneuver type follows a `.takeRoundabout` maneuver. It is only used when `RouteOptions.includesExitRoundaboutManeuver` is set to true.
-     */
-    case exitRoundabout
-    
-    /**
-     The step requires the user to exit a large, named roundabout (traffic circle or rotary).
-     
-     This maneuver type follows a `.takeRotary` maneuver. It is only used when `RouteOptions.includesExitRoundaboutManeuver` is set to true.
-     */
-    case exitRotary
-    
-    /**
-     The step requires the user to respond to a change in travel conditions.
-     
-     This maneuver type may occur for example when driving directions require the user to board a ferry, or when cycling directions require the user to dismount. The step’s transport type and instructions contains important contextual details that should be presented to the user at the maneuver location.
-     
-     Similar changes can occur simultaneously with other maneuvers, such as when the road changes its name at the site of a movable bridge. In such cases, `heedWarning` is suppressed in favor of another maneuver type.
-     */
-    case heedWarning
-    
-    /**
-     The step requires the user to arrive at a waypoint.
-     
-     The distance and expected travel time for this step are set to zero, indicating that the route or route leg is complete. The maneuver direction indicates the side of the road on which the waypoint can be found (or whether it is straight ahead).
-     */
-    case arrive
-    
-    /**
-     The step requires the user to arrive at an intermediate waypoint.
-     
-     This maneuver type is only used by version 4 of the Mapbox Directions API.
-     */
-    case passWaypoint // v4
-    
-    public init?(description: String) {
-        let type: ManeuverType
-        switch description {
-        case "depart":
-            type = .depart
-        case "turn":
-            type = .turn
-        case "continue":
-            type = .continue
-        case "new name":
-            type = .passNameChange
-        case "merge":
-            type = .merge
-        case "on ramp":
-            type = .takeOnRamp
-        case "off ramp":
-            type = .takeOffRamp
-        case "fork":
-            type = .reachFork
-        case "end of road":
-            type = .reachEnd
-        case "use lane":
-            type = .useLane
-        case "rotary":
-            type = .takeRotary
-        case "roundabout":
-            type = .takeRoundabout
-        case "roundabout turn":
-            type = .turnAtRoundabout
-        case "exit roundabout":
-            type = .exitRoundabout
-        case "exit rotary":
-            type = .exitRotary
-        case "notification":
-            type = .heedWarning
-        case "arrive":
-            type = .arrive
-        case "waypoint": // v4
-            type = .passWaypoint
-        default:
-            type = .none
-        }
-        self.init(rawValue: type.rawValue)
-    }
-    
-    public var description: String {
-        switch self {
-        case .none:
-            return "none"
-        case .depart:
-            return "depart"
-        case .turn:
-            return "turn"
-        case .continue:
-            return "continue"
-        case .passNameChange:
-            return "new name"
-        case .merge:
-            return "merge"
-        case .takeOnRamp:
-            return "on ramp"
-        case .takeOffRamp:
-            return "off ramp"
-        case .reachFork:
-            return "fork"
-        case .reachEnd:
-            return "end of road"
-        case .useLane:
-            return "use lane"
-        case .takeRotary:
-            return "rotary"
-        case .takeRoundabout:
-            return "roundabout"
-        case .turnAtRoundabout:
-            return "roundabout turn"
-        case .exitRoundabout:
-            return "exit roundabout"
-        case .exitRotary:
-            return "exit rotary"
-        case .heedWarning:
-            return "notification"
-        case .arrive:
-            return "arrive"
-        case .passWaypoint: // v4
-            return "waypoint"
-        }
-    }
-}
-
-/**
- A `ManeuverDirection` clarifies a `ManeuverType` with directional information. The exact meaning of the maneuver direction for a given step depends on the step’s maneuver type; see the `ManeuverType` documentation for details.
- */
-@objc(MBManeuverDirection)
-public enum ManeuverDirection: Int, CustomStringConvertible, Codable {
-    /**
-     The step does not have a particular maneuver direction associated with it.
-     
-     This maneuver direction is used as a workaround for bridging to Objective-C which does not support nullable enumeration-typed values.
-     */
-    case none
-    
-    /**
-     The maneuver requires a sharp turn to the right.
-     */
-    case sharpRight
-    
-    /**
-     The maneuver requires a turn to the right, a merge to the right, or an exit on the right, or the destination is on the right.
-     */
-    case right
-    
-    /**
-     The maneuver requires a slight turn to the right.
-     */
-    case slightRight
-    
-    /**
-     The maneuver requires no notable change in direction, or the destination is straight ahead.
-     */
-    case straightAhead
-    
-    /**
-     The maneuver requires a slight turn to the left.
-     */
-    case slightLeft
-    
-    /**
-     The maneuver requires a turn to the left, a merge to the left, or an exit on the left, or the destination is on the right.
-     */
-    case left
-    
-    /**
-     The maneuver requires a sharp turn to the left.
-     */
-    case sharpLeft
-    
-    /**
-     The maneuver requires a U-turn when possible.
-     
-     Use the difference between the step’s initial and final headings to distinguish between a U-turn to the left (typical in countries that drive on the right) and a U-turn on the right (typical in countries that drive on the left). If the difference in headings is greater than 180 degrees, the maneuver requires a U-turn to the left. If the difference in headings is less than 180 degrees, the maneuver requires a U-turn to the right.
-     */
-    case uTurn
-    
-    public init?(description: String) {
-        let direction: ManeuverDirection
-        switch description {
-        case "sharp right":
-            direction = .sharpRight
-        case "right":
-            direction = .right
-        case "slight right":
-            direction = .slightRight
-        case "straight":
-            direction = .straightAhead
-        case "slight left":
-            direction = .slightLeft
-        case "left":
-            direction = .left
-        case "sharp left":
-            direction = .sharpLeft
-        case "uturn":
-            direction = .uTurn
-        default:
-            direction = .none
-        }
-        self.init(rawValue: direction.rawValue)
-    }
-    
-    public var description: String {
-        switch self {
-        case .none:
-            return "none"
-        case .sharpRight:
-            return "sharp right"
-        case .right:
-            return "right"
-        case .slightRight:
-            return "slight right"
-        case .straightAhead:
-            return "straight"
-        case .slightLeft:
-            return "slight left"
-        case .left:
-            return "left"
-        case .sharpLeft:
-            return "sharp left"
-        case .uTurn:
-            return "uturn"
-        }
-    }
-}
-
-/**
- A `DrivingSide` indicates which side of the road cars and traffic flow.
- */
-@objc(MBDrivingSide)
-public enum DrivingSide: Int, CustomStringConvertible, Codable {
-    /**
-     Indicates driving occurs on the `left` side.
-     */
-    case left
-    
-    /**
-     Indicates driving occurs on the `right` side.
-     */
-    case right
-    
-    public init?(description: String) {
-        var side: DrivingSide
-        switch description {
-        case "left":
-            side = .left
-        case "right":
-            side = .right
-        default:
-            return nil
-        }
-        
-        self.init(rawValue: side.rawValue)
-    }
-    
-    public var description: String {
-        switch self {
-        case .left:
-            return "left"
-        case .right:
-            return "right"
-        }
-    }
-}
-
-extension String {
-    internal func tagValues(separatedBy separator: String) -> [String] {
-        return components(separatedBy: separator).map { $0.trimmingCharacters(in: .whitespaces) }.filter { !$0.isEmpty }
-    }
-}
-
-/**
- Encapsulates all the information about a road.
- */
-struct Road {
-    let names: [String]?
-    let codes: [String]?
-    let exitCodes: [String]?
-    let destinations: [String]?
-    let destinationCodes: [String]?
-    let rotaryNames: [String]?
-    
-    init(name: String, ref: String?, exits: String?, destination: String?, rotaryName: String?) {
-        var codes: [String]?
-        if !name.isEmpty, let ref = ref {
-            // Mapbox Directions API v5 encodes the ref separately from the name but redundantly includes the ref in the name for backwards compatibility. Remove the ref from the name.
-            let parenthetical = "(\(ref))"
-            if name == ref {
-                self.names = nil
-            } else {
-                self.names = name.replacingOccurrences(of: parenthetical, with: "").tagValues(separatedBy: ";")
-            }
-            codes = ref.tagValues(separatedBy: ";")
-        } else if !name.isEmpty, let codesRange = name.range(of: "\\(.+?\\)$", options: .regularExpression, range: name.startIndex..<name.endIndex) {
-            // Mapbox Directions API v4 encodes the ref inside a parenthetical. Remove the ref from the name.
-            let parenthetical = name[codesRange]
-            if name == ref {
-                self.names = nil
-            } else {
-                self.names = name.replacingOccurrences(of: parenthetical, with: "").tagValues(separatedBy: ";")
-            }
-            codes = parenthetical.trimmingCharacters(in: CharacterSet(charactersIn: "()")).tagValues(separatedBy: ";")
-        } else {
-            self.names = name.isEmpty ? nil : name.tagValues(separatedBy: ";")
-            codes = ref?.tagValues(separatedBy: ";")
-        }
-        
-        // Mapbox Directions API v5 combines the destination’s ref and name.
-        if let destination = destination, destination.contains(": ") {
-            let destinationComponents = destination.components(separatedBy: ": ")
-            self.destinationCodes = destinationComponents.first?.tagValues(separatedBy: ",")
-            self.destinations = destinationComponents.dropFirst().joined(separator: ": ").tagValues(separatedBy: ",")
-        } else {
-            self.destinationCodes = nil
-            self.destinations = destination?.tagValues(separatedBy: ",")
-        }
-        
-        self.exitCodes = exits?.tagValues(separatedBy: ";")
-        self.codes = codes
-        self.rotaryNames = rotaryName?.tagValues(separatedBy: ";")
-    }
-}
 
 /**
  A `RouteStep` object represents a single distinct maneuver along a route and the approach to the next maneuver. The route step object corresponds to a single instruction the user must follow to complete a portion of the route. For example, a step might require the user to turn then follow a road.
@@ -579,11 +33,6 @@ open class RouteStep: NSObject, Codable {
         case phoneticNames = "rotary_pronunciation"
         case transportType
         case rotaryName = "rotary_name"
-    }
-    
-    private enum GeometryCodingKeys: String, CodingKey {
-        case coordinates
-        case type
     }
     
     private enum ManeuverCodingKeys: String, CodingKey {
@@ -644,22 +93,8 @@ open class RouteStep: NSObject, Codable {
             maneuverLocation = CLLocationCoordinate2D()
         }
         
-        // TODO: Move encoding/decoding into ManeuverType
-        
-        if let type = try maneuver.decodeIfPresent(String.self, forKey: .type) {
-        maneuverDirection = ManeuverDirection(description: maneuverDirectionDescription) ?? .none
-        
-            maneuverType = ManeuverType(description: type)
-        } else {
-            maneuverType = nil
-        }
-        
-        // TODO: Move encoding/decoding into ManeuverDirection
-        if let direction = try maneuver.decodeIfPresent(String.self, forKey: .direction) {
-            maneuverDirection = ManeuverDirection(description: direction)
-        } else {
-            maneuverDirection = nil
-        }
+        maneuverType = try maneuver.decodeIfPresent(ManeuverType.self, forKey: .type)
+        maneuverDirection = try maneuver.decodeIfPresent(ManeuverDirection.self, forKey: .direction)
         
         initialHeading = try maneuver.decodeIfPresent(CLLocationDirection.self, forKey: .initialHeading)
         finalHeading = try maneuver.decodeIfPresent(CLLocationDirection.self, forKey: .finalHeading)
@@ -961,39 +396,7 @@ open class RouteStep: NSObject, Codable {
     }
 }
 
-// MARK: Support for Directions API v4
 
-extension ManeuverType {
-    internal init?(v4Description: String) {
-        let description: String
-        switch v4Description {
-        case "bear right", "turn right", "sharp right", "sharp left", "turn left", "bear left", "u-turn":
-            description = "turn"
-        case "enter roundabout":
-            description = "roundabout"
-        default:
-            description = v4Description
-        }
-        self.init(description: description)
-    }
-}
-
-extension ManeuverDirection {
-    internal init?(v4TypeDescription: String) {
-        let description: String
-        switch v4TypeDescription {
-        case "bear right", "bear left":
-            description = v4TypeDescription.replacingOccurrences(of: "bear", with: "slight")
-        case "turn right", "turn left":
-            description = v4TypeDescription.replacingOccurrences(of: "turn ", with: "")
-        case "u-turn":
-            description = "uturn"
-        default:
-            description = v4TypeDescription
-        }
-        self.init(description: description)
-    }
-}
 
 func debugQuickLookURL(illustrating coordinates: [CLLocationCoordinate2D], profileIdentifier: MBDirectionsProfileIdentifier = .automobile) -> URL? {
     guard let accessToken = defaultAccessToken else {
